Shuffle function added

+++ src/main/java/com/kierancaruana/adamusicplayer/objects/Song.java
@@ -3,15 +3,15 @@ package com.kierancaruana.adamusicplayer.objects;
 import javafx.beans.property.*;
 
 public class Song {
-        private final SimpleDoubleProperty trackId = new SimpleDoubleProperty();
+        private final SimpleIntegerProperty trackId = new SimpleIntegerProperty();
         private final StringProperty trackName = new SimpleStringProperty();
         private final SimpleStringProperty trackArtist = new SimpleStringProperty();
         private final SimpleStringProperty trackAlbum = new SimpleStringProperty();
         private String trackFileLocation;
 
         public final double getTrackId(){return this.trackId.get();}
-        public final void setTrackId(double id){trackId.set(id);}
-        public SimpleDoubleProperty trackIdProperty() {return trackId;}
+        public final void setTrackId(int id){trackId.set(id);}
+        public SimpleIntegerProperty trackIdProperty() {return trackId;}
 
         public String getTrackName(){return this.trackName.get();}
         public void setTrackName(final String trackName){this.trackName.set(trackName);}
